requirements.txt
langchain
langchain-groq
langchain-huggingface
langchain-chroma
chromadb
langgraph
sentence-transformers
python-dotenv
pypdf